feat(code-quality): Refactored to turn into an extension
