+++ src/main/java/com/exsoinn/maven/plugin/Test.java
@@ -1,0 +1,8 @@
+package com.exsoinn.maven.plugin;
+
+/**
+ * @author josequijada
+ */
+public class Test {
+
+}
